get_block.BipartiteSBM_fit method

A fct that return blocks attribution or probabilities for each nodes in a Sbm fit from the sbm package.

Usage

# S3 method for BipartiteSBM_fit
get_block(
  x,
  labels = "default",
  node_names = NULL,
  attribution = TRUE,
  proportion = FALSE
)

Arguments

x

Sbm model of class `BipartiteSBM_fit`.

labels

labels for nodes. If it's simple sbm it should be a single character ("default" -> c("nodes")). If sbm is bipartite a named character (names are row and col) ("default" -> c(row = 'row', col = 'col')).

node_names

named list ("row","col"), row is a character vector containing names of nodes in rows, and respectively for columns

attribution

Boolean indicating whether or not the produced tables should contain a block attribution column. This column shows the block in which each nodes is the most likely to be.

proportion

Boolean indicating whether or not the produced tables should contain the probabilities to belong in each blocks. These columns shows for every nodes and every blocks the probabilities that the node belong to the block.

Value

A list containing two data.frames with block attributions and/or proportions one for the row blocks and one for the column blocks

Examples

# my_sbm_bi <- sbm::estimateBipartiteSBM(sbm::fungusTreeNetwork$fungus_tree,
#                                        model = 'bernoulli')
my_sbm_bi <- FungusTreeNetwork$sbmResults$fungus_tree

node_names_bi <- list(
  row = FungusTreeNetwork$networks$fungus_names,
  col = FungusTreeNetwork$networks$tree_names
)

my_blocks_bi <- get_block(my_sbm_bi,
  labels = c(row = "Fungus", col = "Tree"),
  node_names = node_names_bi
)
my_blocks_bi$row
#>                                                                    Nodes_names
#> 1                                                       Amphiporthe leiphaemia
#> 2                                                       Apiognomonia errabunda
#> 3                                                          Apiognomonia veneta
#> 4                                                        Armillaria cepistipes
#> 5                                                           Armillaria gallica
#> 6                                                            Armillaria mellea
#> 7                                                           Armillaria ostoyae
#> 8                                                              Asteroma tiliae
#> 9                                 Biscogniauxia mediterranea var. mediterranea
#> 10                                                          Bjerkandera adusta
#> 11                                                          Blumeriella jaapii
#> 12    Botrytis cinerea (grouped with Botryotinia fuckeliana (anamorphic type))
#> 13                                                     Botryosphaeria dothidea
#> 14                                                    Botryosphaeria stevensii
#> 15                                                           Caliciopsis pinea
#> 16                                                      Cenangium ferruginosum
#> 17                                                            Ceratocystis ips
#> 18                                                          Ceratocystis minor
#> 19                                                        Ceratocystis platani
#> 20                                                    Chondrostereum purpureum
#> 21                                                          Chrysomyxa abietis
#> 22                                            Chrysomyxa ledi var. rhododendri
#> 23                        Coleosporium tussilaginis f.sp. senecionis-silvatici
#> 24                                                           Colpoma quercinum
#> 25                                                   Cristulariella depraedans
#> 26                                                        Cronartium flaccidum
#> 27                                                         Cronartium quercuum
#> 28                                                         Cronartium ribicola
#> 29                                                       Crumenulopsis sororia
#> 30                                                    Cryptodiaporthe castanea
#> 31                                                      Cryptostroma corticale
#> 32                                                     Cryptodiaporthe hystrix
#> 33                                                    Cryphonectria parasitica
#> 34                                                     Cryptodiaporthe populea
#> 35                                                           Cyclaneusma minus
#> 36                                                          Cyclaneusma niveum
#> 37                                                      Cylindrocarpon didymum
#> 38                                                     Daedaleopsis confragosa
#> 39                                                         Delphinella abietis
#> 40                                                           Dicarpella dryina
#> 41                                                           Discella castanea
#> 42                                                  Drepanopeziza punctiformis
#> 43                                                           Entoleuca mammata
#> 44                                                                 Eutypa lata
#> 45                                                          Fistulina hepatica
#> 46                                                           Fomes fomentarius
#> 47                                                         Fomitopsis cytisina
#> 48                                                         Fomitopsis pinicola
#> 49                                                          Fusarium oxysporum
#> 50                                                          Fusicoccum quercus
#> 51                                                         Ganoderma adspersum
#> 52                                                        Ganoderma applanatum
#> 53                                                        Ganoderma resinaceum
#> 54                                                         Gnomonia leptostyla
#> 55                                                       Gremmeniella abietina
#> 56                                                            Gymnopus fusipes
#> 57                                                 Haematonectria haematococca
#> 58                                                       Herpotrichia juniperi
#> 59                                                      Heterobasidion annosum
#> 60                                                        Hypoxylon fragiforme
#> 61                                                      Lachnellula willkommii
#> 62                                                       Laetiporus sulphureus
#> 63                                                    Leptographium lundbergii
#> 64                                                      Leptographium procerum
#> 65                                                        Leucostoma persoonii
#> 66                                                           Lirula macrospora
#> 67                                         Lirula nervisequia var. nervisequia
#> 68                                                      Lophodermium conigenum
#> 69                                                         Lophodermium piceae
#> 70                                                       Lophodermium pinastri
#> 71                                                     Lophodermium seditiosum
#> 72                                                   Melampsora allii-populina
#> 73                                             Melampsorella caryophyllacearum
#> 74                                                 Melampsora laricis-populina
#> 75                                                          Melanconis modonia
#> 76                                                         Melampsora populnea
#> 77                                                         Meripilus giganteus
#> 78                                                    Microsphaera alphitoides
#> 79                                                       Microstroma juglandis
#> 80                                                      Monochaetia monochaeta
#> 81                                                 Mycosphaerella castaneicola
#> 82                                                   Mycosphaerella dearnessii
#> 83                                                    Mycosphaerella microsora
#> 84                                                         Mycosphaerella pini
#> 85                                                         Nectria cinnabarina
#> 86  Nectria coccinea (grouped with Cylindocarpon candidum (teleomorphic type))
#> 87                                                           Nectria ditissima
#> 88                                                          Nectria fuckeliana
#> 89                                                        Nectria veuillotiana
#> 90                                                        Neonectria galligena
#> 91                                                           Ophiostoma piceae
#> 92                                                          Ophiostoma serpens
#> 93                         Ophiostoma ulmi (grouped with Ophiostoma novo-ulmi)
#> 94                                                        Oudemansiella mucida
#> 95                                                          Oxyporus populinus
#> 96                                                      Pestalotiopsis funerea
#> 97                                                         Pezicula cinnamomea
#> 98                                                       Phacidium coniferarum
#> 99                                                   Phaeocryptopus gaeumannii
#> 100                                                      Phaeolus schweinitzii
#> 101                                                       Phellinus chrysoloma
#> 102                                                         Phellinus hartigii
#> 103                                                        Phellinus igniarius
#> 104                                                             Phellinus pini
#> 105                                                         Phellinus robustus
#> 106                                                         Phloeospora aceris
#> 107                                                     Phomopsis juniperivora
#> 108                                                        Phomopsis quercella
#> 109                                                        Phyllosticta aceris
#> 110                                                       Phyllactinia guttata
#> 111                                                        Pleurotus ostreatus
#> 112                                                  Pleuroceras pseudoplatani
#> 113                                    Podosphaera clandestina var. aucupariae
#> 114                                                          Pollaccia radiosa
#> 115                                                        Polyporus squamosus
#> 116                                                    Pseudoinonotus dryadeus
#> 117                                                     Pucciniastrum epilobii
#> 118                                                   Rhabdocline pseudotsugae
#> 119                                                    Rhizosphaera kalkhoffii
#> 120                                                    Rhizosphaera macrospora
#> 121                                                    Rhizosphaera oudemansii
#> 122                                                           Rhizina undulata
#> 123                                                          Rhytisma acerinum
#> 124                                                        Rosellinia necatrix
#> 125                                                              Sarea resinae
#> 126                                                          Sawadaea tulasnei
#> 127                                                      Schizophyllum commune
#> 128                                                        Seiridium cardinale
#> 129                                                      Septotis podophyllina
#> 130                                                        Septoria quercicola
#> 131                                                     Sirococcus strobilinus
#> 132                                                           Sparassis crispa
#> 133                                                        Sphaeropsis sapinea
#> 134                                                    Stegonsporium pyriforme
#> 135                                                           Stereum hirsutum
#> 136                                                     Stereum sanguinolentum
#> 137                                                        Stigmina carpophila
#> 138                                                          Sydowia polyspora
#> 139                                                          Taphrina betulina
#> 140                                                      Taphrina caerulescens
#> 141                                                            Taphrina cerasi
#> 142                                                          Taphrina populina
#> 143                                                     Titaeosporina tremulae
#> 144                                                           Trametes gibbosa
#> 145                                                        Trametes versicolor
#> 146                                                       Truncatella hartigii
#> 147                                                              Valsa ambiens
#> 148                                                               Valsa kunzei
#> 149                                                                Valsa nivea
#> 150                                                              Valsa sordida
#> 151                                                           Venturia fraxini
#> 152                                                        Venturia inaequalis
#> 153                                                       Verticillium dahliae
#> 154                                                      Zythiostroma pinastri
#>       Blocks
#> 1   Fungus_B
#> 2   Fungus_B
#> 3   Fungus_B
#> 4   Fungus_B
#> 5   Fungus_B
#> 6   Fungus_B
#> 7   Fungus_B
#> 8   Fungus_B
#> 9   Fungus_B
#> 10  Fungus_B
#> 11  Fungus_B
#> 12  Fungus_B
#> 13  Fungus_B
#> 14  Fungus_B
#> 15  Fungus_B
#> 16  Fungus_B
#> 17  Fungus_B
#> 18  Fungus_B
#> 19  Fungus_B
#> 20  Fungus_B
#> 21  Fungus_B
#> 22  Fungus_B
#> 23  Fungus_B
#> 24  Fungus_B
#> 25  Fungus_B
#> 26  Fungus_B
#> 27  Fungus_B
#> 28  Fungus_B
#> 29  Fungus_B
#> 30  Fungus_B
#> 31  Fungus_B
#> 32  Fungus_B
#> 33  Fungus_B
#> 34  Fungus_B
#> 35  Fungus_B
#> 36  Fungus_B
#> 37  Fungus_B
#> 38  Fungus_B
#> 39  Fungus_B
#> 40  Fungus_B
#> 41  Fungus_B
#> 42  Fungus_B
#> 43  Fungus_B
#> 44  Fungus_B
#> 45  Fungus_B
#> 46  Fungus_B
#> 47  Fungus_B
#> 48  Fungus_B
#> 49  Fungus_B
#> 50  Fungus_B
#> 51  Fungus_B
#> 52  Fungus_B
#> 53  Fungus_B
#> 54  Fungus_B
#> 55  Fungus_B
#> 56  Fungus_B
#> 57  Fungus_B
#> 58  Fungus_B
#> 59  Fungus_B
#> 60  Fungus_B
#> 61  Fungus_B
#> 62  Fungus_B
#> 63  Fungus_B
#> 64  Fungus_B
#> 65  Fungus_B
#> 66  Fungus_B
#> 67  Fungus_B
#> 68  Fungus_B
#> 69  Fungus_B
#> 70  Fungus_B
#> 71  Fungus_B
#> 72  Fungus_B
#> 73  Fungus_B
#> 74  Fungus_B
#> 75  Fungus_B
#> 76  Fungus_B
#> 77  Fungus_B
#> 78  Fungus_B
#> 79  Fungus_B
#> 80  Fungus_B
#> 81  Fungus_B
#> 82  Fungus_B
#> 83  Fungus_B
#> 84  Fungus_B
#> 85  Fungus_B
#> 86  Fungus_B
#> 87  Fungus_B
#> 88  Fungus_B
#> 89  Fungus_B
#> 90  Fungus_B
#> 91  Fungus_B
#> 92  Fungus_B
#> 93  Fungus_B
#> 94  Fungus_B
#> 95  Fungus_B
#> 96  Fungus_B
#> 97  Fungus_B
#> 98  Fungus_B
#> 99  Fungus_B
#> 100 Fungus_B
#> 101 Fungus_B
#> 102 Fungus_B
#> 103 Fungus_B
#> 104 Fungus_B
#> 105 Fungus_B
#> 106 Fungus_B
#> 107 Fungus_B
#> 108 Fungus_B
#> 109 Fungus_B
#> 110 Fungus_B
#> 111 Fungus_B
#> 112 Fungus_B
#> 113 Fungus_B
#> 114 Fungus_B
#> 115 Fungus_B
#> 116 Fungus_B
#> 117 Fungus_B
#> 118 Fungus_B
#> 119 Fungus_B
#> 120 Fungus_B
#> 121 Fungus_B
#> 122 Fungus_B
#> 123 Fungus_B
#> 124 Fungus_B
#> 125 Fungus_B
#> 126 Fungus_B
#> 127 Fungus_B
#> 128 Fungus_B
#> 129 Fungus_B
#> 130 Fungus_B
#> 131 Fungus_B
#> 132 Fungus_B
#> 133 Fungus_B
#> 134 Fungus_B
#> 135 Fungus_B
#> 136 Fungus_B
#> 137 Fungus_B
#> 138 Fungus_B
#> 139 Fungus_B
#> 140 Fungus_B
#> 141 Fungus_B
#> 142 Fungus_B
#> 143 Fungus_B
#> 144 Fungus_B
#> 145 Fungus_B
#> 146 Fungus_B
#> 147 Fungus_B
#> 148 Fungus_B
#> 149 Fungus_B
#> 150 Fungus_B
#> 151 Fungus_B
#> 152 Fungus_B
#> 153 Fungus_B
#> 154 Fungus_B
my_blocks_bi$col
#>                                                                                             Nodes_names
#> 1                                                                                            Abies alba
#> 2                                                                                         Abies grandis
#> 3                                                                                    Abies nordmanniana
#> 4                                                  Large Maples (Acer platanoides, Acer pseudoplatanus)
#> 5                        Small Maples (Acer campestre, Acer monspessulanum, Acer negundo, Acer opalus) 
#> 6                                                                                       Alnus glutinosa
#> 7                                                       Betulus spp (Betula pendula, Betula pubescens) 
#> 8                                                                                      Carpinus betulus
#> 9                                                                                       Castanea sativa
#> 10                                                        Cedrus spp (Cedrus atlantica, Cedrus libani) 
#> 11                                                                               Cupressus sempervirens
#> 12                                                                                      Fagus silvatica
#> 13                                            Fraxinus spp (Fraxinus angustifolia, Fraxinus excelsior) 
#> 14                                                          Juglans spp (Juglans nigra, Juglans regia) 
#> 15                                                                                        Larix decidua
#> 16                                                                                      Larix kaempferi
#> 17                                                                                        Picea excelsa
#> 18                                                                                     Picea sitchensis
#> 19                                                   Pinus brutia (Pinus brutia, Pinus brutia eldarica)
#> 20                                                                                         Pinus cembra
#> 21                                                                                     Pinus halepensis
#> 22                                                                                  Pinus nigra laricio
#> 23                                                                                    Pinus nigra nigra
#> 24                                                                                       Pinus pinaster
#> 25                                                                                          Pinus pinea
#> 26                                                                                        Pinus radiata
#> 27                                                                                        Pinus strobus
#> 28                                                                                     Pinus sylvestris
#> 29                                                                                          Pinus taeda
#> 30                                                                                       Pinus uncinata
#> 31                                                                                     Platanus hybrida
#> 32 Cultivated Poplars (Populus trichocarpa, P. canescens, P.alba, P.nigra and their cultivated hybrids)
#> 33                                                                                      Populus tremula
#> 34                                                                                         Prunus avium
#> 35                                                                                Pseudotsuga menziesii
#> 36                                                                                         Quercus ilex
#> 37                                                                                      Quercus petraea
#> 38                                                                                    Quercus pubescens
#> 39                                                                                    Quercus pyrenaica
#> 40                                                                                        Quercus robur
#> 41                                                                                        Quercus rubra
#> 42                                                                                        Quercus suber
#> 43                                                                                          Sorbus aria
#> 44                                                                                     Sorbus aucuparia
#> 45                                                                                     Sorbus domestica
#> 46                                                                                    Sorbus torminalis
#> 47                                                                                        Taxus baccata
#> 48                                                                                        Thuja plicata
#> 49                                                       Tilia spp (Tilia platiphyllos, Tilia cordata) 
#> 50                                                                                   Tsuga heterophylla
#> 51                                                 Ulmus spp (Ulmus minor, Ulmus laevis, Ulmus glabra) 
#>    Blocks
#> 1  Tree_B
#> 2  Tree_B
#> 3  Tree_B
#> 4  Tree_B
#> 5  Tree_B
#> 6  Tree_B
#> 7  Tree_B
#> 8  Tree_B
#> 9  Tree_B
#> 10 Tree_B
#> 11 Tree_B
#> 12 Tree_B
#> 13 Tree_B
#> 14 Tree_B
#> 15 Tree_B
#> 16 Tree_B
#> 17 Tree_B
#> 18 Tree_B
#> 19 Tree_B
#> 20 Tree_B
#> 21 Tree_B
#> 22 Tree_B
#> 23 Tree_B
#> 24 Tree_B
#> 25 Tree_B
#> 26 Tree_B
#> 27 Tree_B
#> 28 Tree_B
#> 29 Tree_B
#> 30 Tree_B
#> 31 Tree_B
#> 32 Tree_B
#> 33 Tree_B
#> 34 Tree_B
#> 35 Tree_B
#> 36 Tree_B
#> 37 Tree_B
#> 38 Tree_B
#> 39 Tree_B
#> 40 Tree_B
#> 41 Tree_B
#> 42 Tree_B
#> 43 Tree_B
#> 44 Tree_B
#> 45 Tree_B
#> 46 Tree_B
#> 47 Tree_B
#> 48 Tree_B
#> 49 Tree_B
#> 50 Tree_B
#> 51 Tree_B