get_block.SimpleSBM_fit method

A fct that return blocks attribution or probabilities for each nodes in a Sbm fit from the sbm package.

Usage

# S3 method for SimpleSBM_fit
get_block(
  x,
  labels = "default",
  node_names = NULL,
  attribution = TRUE,
  proportion = FALSE
)

Arguments

x

Sbm model of class `SimpleSBM_fit`.

labels

labels for nodes. If it's simple sbm it should be a single character ("default" -> c("nodes")). If sbm is bipartite a named character (names are row and col) ("default" -> c(row = 'row', col = 'col')).

node_names

character: node names

attribution

Boolean indicating whether or not the produced tables should contain a block attribution column. This column shows the block in which each nodes is the most likely to be.

proportion

Boolean indicating whether or not the produced tables should contain the probabilities to belong in each blocks. These columns shows for every nodes and every blocks the probabilities that the node belong to the block.

Value

A data.frame with block attributions and/or proportions

Examples

# my_sbm_uni <- sbm::estimateSimpleSBM(sbm::fungusTreeNetwork$tree_tree,
#                                      model = "poisson")
my_sbm_uni <- FungusTreeNetwork$sbmResults$tree_tree

node_names_uni <- list(FungusTreeNetwork$networks$tree_names)

my_blocks_uni <- get_block(my_sbm_uni,
  labels = c("Tree"),
  node_names = node_names_uni
)
my_blocks_uni
#>                                                                                             Nodes_names
#> 1                                                                                            Abies alba
#> 2                                                                                         Abies grandis
#> 3                                                                                    Abies nordmanniana
#> 4                                                  Large Maples (Acer platanoides, Acer pseudoplatanus)
#> 5                        Small Maples (Acer campestre, Acer monspessulanum, Acer negundo, Acer opalus) 
#> 6                                                                                       Alnus glutinosa
#> 7                                                       Betulus spp (Betula pendula, Betula pubescens) 
#> 8                                                                                      Carpinus betulus
#> 9                                                                                       Castanea sativa
#> 10                                                        Cedrus spp (Cedrus atlantica, Cedrus libani) 
#> 11                                                                               Cupressus sempervirens
#> 12                                                                                      Fagus silvatica
#> 13                                            Fraxinus spp (Fraxinus angustifolia, Fraxinus excelsior) 
#> 14                                                          Juglans spp (Juglans nigra, Juglans regia) 
#> 15                                                                                        Larix decidua
#> 16                                                                                      Larix kaempferi
#> 17                                                                                        Picea excelsa
#> 18                                                                                     Picea sitchensis
#> 19                                                   Pinus brutia (Pinus brutia, Pinus brutia eldarica)
#> 20                                                                                         Pinus cembra
#> 21                                                                                     Pinus halepensis
#> 22                                                                                  Pinus nigra laricio
#> 23                                                                                    Pinus nigra nigra
#> 24                                                                                       Pinus pinaster
#> 25                                                                                          Pinus pinea
#> 26                                                                                        Pinus radiata
#> 27                                                                                        Pinus strobus
#> 28                                                                                     Pinus sylvestris
#> 29                                                                                          Pinus taeda
#> 30                                                                                       Pinus uncinata
#> 31                                                                                     Platanus hybrida
#> 32 Cultivated Poplars (Populus trichocarpa, P. canescens, P.alba, P.nigra and their cultivated hybrids)
#> 33                                                                                      Populus tremula
#> 34                                                                                         Prunus avium
#> 35                                                                                Pseudotsuga menziesii
#> 36                                                                                         Quercus ilex
#> 37                                                                                      Quercus petraea
#> 38                                                                                    Quercus pubescens
#> 39                                                                                    Quercus pyrenaica
#> 40                                                                                        Quercus robur
#> 41                                                                                        Quercus rubra
#> 42                                                                                        Quercus suber
#> 43                                                                                          Sorbus aria
#> 44                                                                                     Sorbus aucuparia
#> 45                                                                                     Sorbus domestica
#> 46                                                                                    Sorbus torminalis
#> 47                                                                                        Taxus baccata
#> 48                                                                                        Thuja plicata
#> 49                                                       Tilia spp (Tilia platiphyllos, Tilia cordata) 
#> 50                                                                                   Tsuga heterophylla
#> 51                                                 Ulmus spp (Ulmus minor, Ulmus laevis, Ulmus glabra) 
#>     Blocks
#> 1  Tree_B1
#> 2  Tree_B1
#> 3  Tree_B3
#> 4  Tree_B2
#> 5  Tree_B4
#> 6  Tree_B4
#> 7  Tree_B6
#> 8  Tree_B6
#> 9  Tree_B4
#> 10 Tree_B1
#> 11 Tree_B5
#> 12 Tree_B2
#> 13 Tree_B2
#> 14 Tree_B2
#> 15 Tree_B1
#> 16 Tree_B3
#> 17 Tree_B1
#> 18 Tree_B3
#> 19 Tree_B5
#> 20 Tree_B5
#> 21 Tree_B3
#> 22 Tree_B1
#> 23 Tree_B3
#> 24 Tree_B1
#> 25 Tree_B5
#> 26 Tree_B5
#> 27 Tree_B3
#> 28 Tree_B1
#> 29 Tree_B5
#> 30 Tree_B3
#> 31 Tree_B5
#> 32 Tree_B2
#> 33 Tree_B6
#> 34 Tree_B2
#> 35 Tree_B1
#> 36 Tree_B4
#> 37 Tree_B2
#> 38 Tree_B4
#> 39 Tree_B6
#> 40 Tree_B2
#> 41 Tree_B2
#> 42 Tree_B4
#> 43 Tree_B4
#> 44 Tree_B6
#> 45 Tree_B6
#> 46 Tree_B2
#> 47 Tree_B6
#> 48 Tree_B6
#> 49 Tree_B4
#> 50 Tree_B5
#> 51 Tree_B6